+++ lib/paper_trail/version_concern.rb
@@ -196,7 +196,15 @@ module PaperTrail
           inheritance_column_name = item_type.constantize.inheritance_column
           class_name = attrs[inheritance_column_name].blank? ? item_type : attrs[inheritance_column_name]
           klass = class_name.constantize
-          model = klass.new
+          # the `dup` option always returns a new object, otherwise we should attempt
+          # to look for the item outside of default scope(s)
+          if options[:dup] || (_item = klass.unscoped.find_by_id(item_id)).nil?
+            model = klass.new
+          else
+            model = _item
+            # Look for attributes that exist in the model and not in this version. These attributes should be set to nil.
+            (model.attribute_names - attrs.keys).each { |k| attrs[k] = nil }
+          end
         end
 
         if PaperTrail.serialized_attributes?