+++ test/dummy/app/models/boolit.rb
@@ -1,0 +1,4 @@
+class Boolit < ActiveRecord::Base
+  default_scope { where(:scoped => true) }
+  has_paper_trail
+end